updated schema and gifs for readme

+++ schema.sql
@@ -8,7 +8,7 @@ CREATE TABLE products (
     department_name VARCHAR(100),
     price DECIMAL(10,2),
     stock_quantity INTEGER(11),
-    product_sales INTEGER(11),
+    product_sales DECIMAL(10,2),
     PRIMARY KEY (item_id)
 );
